# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Apache Airavata HTTP API
  version: "1.0.0"
  description: |
    The REST API served by `airavata-server` (see `internal/server/server.go`).

    All request and response bodies are JSON (`Content-Type: application/json`).

    ## Authentication

    Authentication is a bearer token validated by CILogon introspection, carried as
    `Authorization: Bearer <token>`. A request without a token is anonymous rather
    than rejected at the door: each operation decides for itself whether an anonymous
    caller is acceptable. A token that is present but invalid is always `401`, with
    `WWW-Authenticate: Bearer error="invalid_token"`. If the identity provider cannot
    be reached the call fails `502`, because that is not the caller's fault.

    ## Authorization

    Two schemes are in play, depending on the resource:

    - **Platform roles** (`ADMIN`, `SUPER_ADMIN`) guard the shared catalogue —
      Slurm clusters, partitions, application templates and batch deployments.
      Reads of the catalogue are open to anyone, including anonymous callers;
      writes require a role.
    - **Ownership and sharing** guards the per-user resources — SSH keys, groups,
      cluster configs, SCP data storages, data products and processes. Ownership
      comes from the token and is never carried in a request body, so nothing can be
      created on someone else's behalf or handed over by editing it. None of these
      are readable anonymously.

...

...

...

...
    state.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/scp-data-storages:
    get:
      tags: [SCP Data Storages]
      summary: List all SCP data storages
      description: Requires `ADMIN` or `SUPER_ADMIN` authority.
      operationId: listSCPDataStorages
      responses:
        "200":
          description: Every storage.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [SCP Data Storages]
      summary: Register an SCP data storage
      description: |
        The owner comes from the token. The key must be one the caller owns. An
        omitted `port` means 22.
      operationId: createSCPDataStorage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SCPDataStorageRequest" }
      responses:
        "201":
          description: The registered storage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/me:
    get:
      tags: [SCP Data Storages]
      summary: List the caller's SCP data storages
      operationId: listMySCPDataStorages
      responses:
        "200":
          description: The storages the caller owns.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/shared-with-me:
    get:
      tags: [SCP Data Storages]
      summary: List SCP data storages shared with the caller
      description: |
        Each entry carries `permission`, the caller's effective standing. `WRITE`
        implies `READ`, and where several shares reach the same caller the strongest
        applies.
      operationId: listSCPDataStoragesSharedWithMe
      responses:
        "200":
          description: The storages shared with the caller.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/{dataStorageId}:
    parameters:
      - $ref: "#/components/parameters/DataStorageId"
    get:
      tags: [SCP Data Storages]
      summary: Read an SCP data storage
      description: Requires a `READ` standing. The key is inlined as its safe summary.
      operationId: getSCPDataStorage
      responses:
        "200":
          description: The storage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [SCP Data Storages]
      summary: Update an SCP data storage
      description: Requires a `WRITE` standing.
      operationId: updateSCPDataStorage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SCPDataStorageRequest" }
      responses:
        "200":
          description: The updated storage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SCPDataStorageResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [SCP Data Storages]
      summary: Delete an SCP data storage
      description: Owner or admin.
      operationId: deleteSCPDataStorage
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/{dataStorageId}/group-shares:
    parameters:
      - $ref: "#/components/parameters/DataStorageId"
    get:
      tags: [SCP Data Storages]
      summary: List a storage's group shares
      description: Owner or admin.
      operationId: listSCPDataStorageGroupShares
      responses:
        "200":
          description: The group shares on this storage.
          content:
            application/json:
              schema:
                type: array
                items:
                  {
                    $ref: "#/components/schemas/SCPDataStorageGroupSharingResponse",
                  }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [SCP Data Storages]
      summary: Share a storage with a group
      description: An omitted `permission` grants `READ`.
      operationId: shareSCPDataStorageWithGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/SCPDataStorageGroupSharingRequest" }
      responses:
        "201":
          description: The created share.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/SCPDataStorageGroupSharingResponse",
                }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/{dataStorageId}/group-shares/{sharingId}:
    parameters:
      - $ref: "#/components/parameters/DataStorageId"
      - $ref: "#/components/parameters/SharingId"
    put:
      tags: [SCP Data Storages]
      summary: Change what a group share grants
      operationId: updateSCPDataStorageGroupShare
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SCPDataStorageSharingUpdate" }
      responses:
        "200":
          description: The updated share.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/SCPDataStorageGroupSharingResponse",
                }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [SCP Data Storages]
      summary: Revoke a group share
      operationId: revokeSCPDataStorageGroupShare
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/{dataStorageId}/user-shares:
    parameters:
      - $ref: "#/components/parameters/DataStorageId"
    get:
      tags: [SCP Data Storages]
      summary: List a storage's user shares
      description: Owner or admin.
      operationId: listSCPDataStorageUserShares
      responses:
        "200":
          description: The user shares on this storage.
          content:
            application/json:
              schema:
                type: array
                items:
                  {
                    $ref: "#/components/schemas/SCPDataStorageUserSharingResponse",
                  }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [SCP Data Storages]
      summary: Share a storage with one user
      description: An omitted `permission` grants `READ`.
      operationId: shareSCPDataStorageWithUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/SCPDataStorageUserSharingRequest" }
      responses:
        "201":
          description: The created share.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/SCPDataStorageUserSharingResponse",
                }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scp-data-storages/{dataStorageId}/user-shares/{sharingId}:
    parameters:
      - $ref: "#/components/parameters/DataStorageId"
      - $ref: "#/components/parameters/SharingId"
    put:
      tags: [SCP Data Storages]
      summary: Change what a user share grants
      operationId: updateSCPDataStorageUserShare
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SCPDataStorageSharingUpdate" }
      responses:
        "200":
          description: The updated share.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/SCPDataStorageUserSharingResponse",
                }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [SCP Data Storages]
      summary: Revoke a user share
      operationId: revokeSCPDataStorageUserShare
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }

  # --------------------------------------------------------- Data products
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A CILogon access token, validated by RFC 7662 introspection. The bootstrap
        root token is also accepted; see INSTALL.md.
  parameters:
...
    DataStorageId:
      name: dataStorageId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    SharingId:
      name: sharingId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NoContent:
      description: Deleted. No body.
    BadRequest:
      description: |
        Malformed JSON, or a body that failed its constraints. Validation failures
        carry `fieldErrors`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            validation:
              summary: Validation failed
              value:
                status: 400
                error: Bad Request
                message: Validation failed
                fieldErrors:
                  - field: clusterName
                    message: Cluster name cannot be blank
    Unauthorized:
      description: |
        No token where one is required, or a token that is present but unusable. An
        invalid token also carries `WWW-Authenticate: Bearer error="invalid_token"`.
      headers:
        WWW-Authenticate:
          description: Present when the token was rejected.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated, but lacking the authority for this call.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such record — or, for a group, a record the caller has no standing to know
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            status: 404
            error: Not Found
            message: "Slurm cluster not found: c1d2e3f4-5678-4abc-9def-0123456789ab"
    Conflict:
      description: |
        The request collides with existing state — deleting an SSH key still in use,
        or a template that still has deployments.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: |
        The identity provider could not be reached to validate the token. Not the
        caller's fault, so it is not reported as an auth failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: |
        An unexpected server failure. The message is always `"Internal server error"`;
        the detail is logged rather than returned, since it may name internal state.

        A `502` is also possible on any authenticated call if the identity provider is
        unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    FieldError:
      type: object
      description: One failed validation constraint.
      properties:
        field:
          type: string
          description: |
            The JSON path of the offending field, including indexes and nesting.
          examples:
            ["inputs[0].inputType", "defaultBatchJobConfig.wallTimeMinutes"]
        message: { type: string }
      required: [field, message]
    Error:
      type: object
      description: The envelope every failure response uses.
      properties:
        status:
          type: integer
          description: The HTTP status, repeated in the body.
        error:
          type: string
          description: The standard reason phrase for `status`.
        message:
          type: string
          description: Human-readable detail.
        fieldErrors:
          type: array
          description: |
            Present only on request-body validation failures, listing every
            constraint that failed so a client can report all of them at once.
          items: { $ref: "#/components/schemas/FieldError" }
      required: [status, error, message]

    # ----------------------------------------------------------- Enums
...
    DataStoragePermission:
      type: string
      enum: [READ, WRITE]
...
    SSHKeyResponse:
      type: object
      description: |
        The read model for a key. It has no private key or passphrase field at all —
        the secrets cannot leak through this endpoint because there is nowhere for
        them to go.
      properties:
        sshKeyId: { type: string, format: uuid }
        sshKeyName: { type: string }
        publicKey: { type: string }
        ownerId: { type: string }
      required: [sshKeyId, sshKeyName, publicKey, ownerId]

    # -------------------------------------------------------- Compute
...
    SCPDataStorageRequest:
      type: object
      description: |
        The create/update payload for a storage. There is no owner field.
      properties:
        dataName: { type: string }
        hostName: { type: string }
        port:
          type: [integer, "null"]
          minimum: 1
          maximum: 65535
          default: 22
        loginUser: { type: string }
        sshKeyId:
          type: string
          format: uuid
          description: Must be a key the caller owns.
      required: [dataName, hostName, loginUser, sshKeyId]
    SCPDataStorageResponse:
      type: object
      properties:
        dataId: { type: string, format: uuid }
        dataName: { type: [string, "null"] }
        ownerId: { type: [string, "null"] }
        hostName: { type: [string, "null"] }
        port: { type: [integer, "null"] }
        loginUser: { type: [string, "null"] }
        sshKeyId: { type: [string, "null"], format: uuid }
        sshKey:
          oneOf:
            - { $ref: "#/components/schemas/SSHKeyResponse" }
            - { type: "null" }
        permission:
          type: string
          description: |
            The caller's effective standing, present only on the listings that compute
            it (`/shared-with-me`).
          enum: [READ, WRITE]
      required: [dataId]
    SCPDataStorageGroupSharingRequest:
      type: object
      properties:
        groupId: { type: string }
        permission:
          oneOf:
            - { $ref: "#/components/schemas/DataStoragePermission" }
            - { type: "null" }
      required: [groupId]
    SCPDataStorageUserSharingRequest:
      type: object
      properties:
        userId: { type: string }
        permission:
          oneOf:
            - { $ref: "#/components/schemas/DataStoragePermission" }
            - { type: "null" }
      required: [userId]
    SCPDataStorageSharingUpdate:
      type: object
      properties:
        permission: { $ref: "#/components/schemas/DataStoragePermission" }
      required: [permission]
    SCPDataStorageGroupSharingResponse:
      type: object
      properties:
        dataStorageGroupSharingId: { type: string, format: uuid }
        dataStorageId: { type: [string, "null"], format: uuid }
        groupId: { type: [string, "null"] }
        permission:
          oneOf:
            - { $ref: "#/components/schemas/DataStoragePermission" }
            - { type: "null" }
      required: [dataStorageGroupSharingId]
    SCPDataStorageUserSharingResponse:
      type: object
      properties:
        dataStorageUserSharingId: { type: string, format: uuid }
        dataStorageId: { type: [string, "null"], format: uuid }
        userId: { type: [string, "null"] }
        permission:
          oneOf:
            - { $ref: "#/components/schemas/DataStoragePermission" }
            - { type: "null" }
      required: [dataStorageUserSharingId]
...